Add unit test and remove cmake generated folder for demo

+++ cfgtrace/engine/engine.h
@@ -22,14 +22,10 @@ private:
 
 public:
     engine(HANDLE file_mapping);
-    engine() : memory(nullptr)
-    {
-    }
+    engine() : memory(nullptr) {}
     engine &operator=(engine other);
     ~engine() = default;
-    PluginLayer *
-    plugin_interface(char *pluginname, size_t layer, PluginLayer **layers) const
-      noexcept;
+    PluginLayer *plugin_interface(char *pluginname, size_t layer, PluginLayer **layers) const noexcept;
     char *log_name() const noexcept;
     char *plugin_path() const noexcept;
     uint8_t *context() const noexcept;
@@ -39,9 +35,8 @@ public:
     size_t process_stacktop() const;
     uint8_t *cfg_memory_region() const noexcept;
     size_t cfg_memory_region_size() const noexcept;
-    int *cfg_iteration() const;
-    int *cfg_size() const;
+    size_t *cfg_iteration() const;
+    size_t *cfg_size() const;
     uint8_t *cfg_serialize_memory_region() const;
 };
-
 }; // namespace engine
